Added support for a 'mirrored' default storage pool for more convinient usage with virt-manager (ON-665)

+++ opennode/cli/actions/storage.py
@@ -1,9 +1,10 @@
 
 import os
 import re
+from xml.dom.minidom import parseString
 
 from opennode.cli.config import c, cs
-from opennode.cli.actions.utils import del_folder, execute, mkdir_p
+from opennode.cli.actions.utils import del_folder, execute, mkdir_p, CommandException
 
 
 __all__ = ['list_pools', 'set_default_pool', 'prepare_storage_pool']
@@ -13,33 +14,71 @@ def list_pools():
     """List existing storage pools"""
     pool_params = []
     try:
-        pools = execute("virsh 'pool-list' | tail -n+3 |head -n-1").splitlines()
+        pools = execute("virsh 'pool-list' | tail -n+3 |head -n-1 | egrep -v '^default-iso |^default '").splitlines()
         for p in pools:
-            p = re.sub("\s+" , " ", p.strip())
+            p = re.sub("\s+", " ", p.strip())
             pool_params.append(p.split(' '))
     except Exception, e:
-        print "Unable to list storage pools: %s" %e
+        print "Unable to list storage pools: %s" % e
     return pool_params
+
+
+def is_default_pool_modified():
+    """Check if there were any modifications done by a user to the default pool"""
+    try:
+        res = execute("virsh 'pool-dumpxml default'")
+        defined_path = parseString(res).getElementsByTagName('path')[0].lastChild.nodeValue
+        current_path = os.path.join(c('general', 'storage-endpoint'), c('general', 'default-storage-pool'), 'images')
+        return str(defined_path) != current_path
+    except CommandException:
+        return False  # pool is undefined or we are not sure -> so, assume it's all good
+
 
 def set_default_pool(name):
     """Set default storage pool"""
+    if name == 'default':
+        raise CommandException('Cannot set pool name to a reserved "default".')
+
+    # clean up default pool
+    if not is_default_pool_modified():
+        for pool_name in ['default', 'default-iso']:
+            try:
+                execute("virsh 'pool-destroy %s'" % pool_name)
+                execute("virsh 'pool-undefine %s'" % pool_name)
+            except CommandException:
+                pass  # it's ok for these commands to fail if the pool is undefined
+        endpoint = c('general', 'storage-endpoint')
+        # create default image pool
+        paths = {'default': 'images',
+                 'default-iso': 'iso'}
+        if len(name.strip()) > 0:
+            for default_pool_name, item_path in paths.items():
+                execute("virsh 'pool-define-as --name %s --type dir --target %s/%s/%s'" % (default_pool_name,
+                                                                                           endpoint, name, item_path))
+                execute("virsh 'pool-autostart %s'" % default_pool_name)
+                execute("virsh 'pool-build %s'" % default_pool_name)
+                execute("virsh 'pool-start %s'" % default_pool_name)
+    # finally set a pointer in the configuration file
     cs('general', 'default-storage-pool', name)
+
 
 def get_default_pool():
     """Return name of the storage pool to use by default. Or None if not configured"""
     name = c('general', 'default-storage-pool')
     return None if name is None or name == '' or name == 'None' else name
 
+
 def delete_pool(pool_name):
     """Delete a storage pool"""
     try:
-        execute("virsh 'pool-destroy %s'" %pool_name)
-        execute("virsh 'pool-undefine %s'" %pool_name)
+        execute("virsh 'pool-destroy %s'" % pool_name)
+        execute("virsh 'pool-undefine %s'" % pool_name)
         del_folder(os.path.join(c('general', 'storage-endpoint'), pool_name))
         if pool_name == c('general', 'default-storage-pool'):
             set_default_pool('')
     except Exception, e:
         print "Failed to delete pool %s: %s" % (pool_name, e)
+
 
 def add_pool(pool_name, careful=True):
     """Add a new pool_name"""
@@ -56,6 +95,7 @@ def add_pool(pool_name, careful=True):
         execute("virsh 'pool-autostart %s'" %pool_name)
     except Exception, e:
         print "Failed to create a new pool: %s" %e
+
 
 def prepare_storage_pool(storage_pool=get_default_pool()):
     """Assures that storage pool has the correct folder structure"""
